https://jira.ccpgames.com/browse/TQ-131403
Fix conformance tweaks for VS2010. Versions of VS earlier than 2015 don't support the standard 'z' printf flag, so we just use long longs instead.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/BlueExposure/": change = 1462450]

+++ BlueClasses.cpp
@@ -647,22 +647,22 @@ std::string BlueClasses::GetPersistedClassesReport()
 				{
                 case Be::FLOATARRAY:
 					{
-						size_t memberCount = entry->GetFloatArraySize();
-						sprintf_s( buffer, "    -   %zd\n", memberCount );
+						long long memberCount = static_cast<long long>( entry->GetFloatArraySize() );
+						sprintf_s( buffer, "    -   %lld\n", memberCount );
 						result += buffer;
 					}
 					break;
                 case Be::DOUBLEARRAY:
 					{
-						size_t memberCount = entry->GetDoubleArraySize();
-						sprintf_s( buffer, "    -   %zd\n", memberCount );
+						long long memberCount = static_cast<long long>( entry->GetDoubleArraySize() );
+						sprintf_s( buffer, "    -   %lld\n", memberCount );
 						result += buffer;
 					}
 					break;
                 case Be::INTARRAY:
 					{
-						size_t memberCount = entry->GetIntArraySize();
-						sprintf_s( buffer, "    -   %zd\n", memberCount );
+						long long memberCount = static_cast<long long>( entry->GetIntArraySize() );
+						sprintf_s( buffer, "    -   %lld\n", memberCount );
 						result += buffer;
 					}
 					break;
